requirements.txt
python-telegram-bot==13.15
Telethon==1.29.3
Flask==2.3.3
pymongo==4.7.0
python-dotenv==1.0.1